Refactor for better separation of concerns
